my visit to Milwaukee
These times have been anything but good. People have abused my trust in them and then those who claimed to care about me manipulated their conditions on me and I was hated by them that are just as flawed as me without the chance to voice my own explanation for the way things seemed or were cause it was not the truth they were saying but you know God got something for us all and  some of us are gonna be blessed due to the sincere transparency of whom we were desiring nothing evil on no one but weaker than we thought by not being able to just turn the other cheek like the Lord Jesus Christ. I am still pushing forward with my music publishing business so don't count your luck haters my words are gonna state the truth of all that I've been a witness to so help me God.